requirements.txt
streamlit
pandas
numpy
matplotlib
mistralai
fpdf2
